Left & Right
Since I was online, I thought I'd input the left and right script. It works, but now the players are working with very little gravity and are falling slowly and, if the up button is pressed enough, they can fly. The leftright script is still in the project, it's just not attached to anything.